feat: divided root options from connection options
